replace original component
