--- src/apps/EventDetail/EventDetail.css
@@ -1,294 +1,0 @@
-.breadcrumbs-container {
-  position: sticky;
-  top: 60px;
-  z-index: 2;
-}
-
-.event-detail {
-  box-sizing: border-box;
-}
-
-.event-detail .set-info-bar {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
-
-.event-detail .set-info-bar .set-info-bar-right {
-  display: flex;
-  gap: 8px;
-  width: 60%;
-  align-items: center;
-  justify-content: flex-end;
-}
-
-.event-detail .set-info-bar .set-info-bar-right .select-trigger {
-  flex-grow: 0;
-}
-
-.event-detail .add-set-button svg {
-  margin: 0;
-}
-
-.delete-confirmation-dialog {
-  z-index: 100;
-}
-
-.event-detail .av-file-selection {
-  display: flex;
-  gap: 10px;
-  align-items: center;
-  margin-bottom: 20px;
-}
-
-.event-detail .av-file-selection .select-trigger {
-  box-shadow: none;
-  width: auto;
-}
-
-.event-detail .av-file-selection .av-file-label {
-  font-weight: 600;
-}
-
-.event-detail .event-detail-label {
-  display: flex;
-  align-items: center;
-  gap: 10px;
-}
-
-.event-detail .event-detail-label h2 {
-  font-size: 24px;
-  margin: 0;
-}
-
-.event-detail .event-detail-label svg {
-  height: 24px;
-  width: 24px;
-}
-
-.event-detail .event-detail-options {
-  display: flex;
-  gap: 10px;
-  height: 38px;
-}
-
-.event-detail .event-detail-options .event-detail-button {
-  padding: 9px 16px 9px 16px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  height: 100%;
-}
-
-.event-detail .event-detail-options .event-detail-button svg {
-  height: 16px;
-  width: 16px;
-}
-
-.event-detail .event-detail-top-bar {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.event-detail-options a {
-  padding: 0;
-  margin: 0;
-}
-
-.event-detail-options a:hover {
-  text-decoration: none;
-}
-
-.event-detail .event-detail-floating-header {
-  position: sticky;
-  top: 120px;
-  display: block;
-  z-index: 2;
-  background-color: var(--gray-100);
-  padding-top: 10px;
-}
-
-.event-detail .rt-TableRootTable {
-  background: #fff;
-  border-radius: 0;
-  table-layout: fixed;
-}
-
-.audio-event-detail {
-  padding-bottom: 60px;
-}
-
-.audio-event-detail .event-detail-table-header {
-  background: #fff;
-  border: 1px solid var(--gray-200);
-  padding: 0 16px;
-}
-
-.audio-event-detail .event-detail-table-header .header-buttons {
-  display: flex;
-  gap: 12px;
-}
-
-.event-detail .tag-cell-container,
-.event-detail .invisible-tag-holder {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 10px;
-}
-
-.event-detail .tag-cell-container .tag-cell-overflow {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  border-radius: 16px;
-  background: var(--gray-200);
-  padding: 4px 10px;
-  font-weight: 700;
-}
-
-.event-detail .tag-item {
-  border-radius: 50px;
-  padding: 6px 12px 6px 12px;
-  font-size: 14px;
-}
-
-.event-detail .formic-form-field {
-  display: flex;
-  align-items: center;
-  position: relative;
-}
-
-.event-detail .formic-form-text {
-  width: 260px;
-}
-
-.event-detail .formic-form-field svg {
-  position: absolute;
-  right: 10px;
-  height: 16px;
-  width: 16px;
-}
-
-.invisible-tag-holder {
-  position: absolute;
-  z-index: -1;
-  opacity: 0;
-}
-
-.audio-event-detail .event-detail-table-container {
-  margin: 28px 0;
-}
-
-.video-event-detail {
-  display: grid;
-  grid-template-columns: 4fr 2fr;
-  background: #fff;
-  box-sizing: border-box;
-  margin-top: 60px;
-  margin-bottom: 60px;
-}
-
-.video-event-detail .formic-form-field {
-  width: 100%;
-}
-
-.video-event-detail .formic-form-field input {
-  width: 100%;
-}
-
-.video-event-detail .video-pane {
-  max-height: calc(100vh - 172px);
-  overflow-y: scroll;
-  padding: 26px;
-}
-
-.video-event-detail .table-container {
-  max-height: calc(100% - 230px);
-  overflow-y: auto;
-  box-sizing: border-box;
-  padding: 0 26px;
-}
-
-.video-event-detail .event-detail-table-container {
-  margin: 20px 0;
-}
-
-.video-event-detail .searchbox {
-  width: 100%;
-  margin: 10px 0;
-  box-sizing: border-box;
-}
-
-.video-event-detail .annotations-pane {
-  background: var(--gray-100);
-  max-height: calc(100vh - 120px);
-}
-
-.video-event-detail .annotations-pane .annotations-pane-header {
-  padding: 0 26px;
-  box-shadow: 0px 1px 2px 0px #00000026;
-}
-
-.video-event-detail .timestamp-cell {
-  width: 100px;
-}
-
-.video-event-detail .timestamp-cell {
-  text-wrap: wrap;
-}
-
-.video-event-detail .select-container {
-  display: flex;
-  align-items: center;
-  margin-top: 10px;
-  flex-flow: row-reverse;
-  gap: 10px;
-}
-
-.video-event-detail .select-trigger {
-  flex-grow: 1;
-}
-
-.video-event-detail .event-detail-table-header .header-buttons {
-  display: flex;
-  gap: 12px;
-  padding: 0;
-}
-
-.video-event-detail .event-detail-table-header .header-buttons button {
-  margin: 0;
-}
-
-.video-event-detail .tag-cell-container {
-  margin: 10px 0;
-}
-
-.video-event-detail .set-picker {
-  margin: 20px 0;
-}
-
-.video-event-detail .set-picker h3 {
-  margin: 0;
-  height: 32px;
-  align-content: center;
-}
-
-.video-event-detail .add-set-button {
-  box-shadow: none;
-  color: var(--gray-900) !important;
-  border-color: var(--gray-900);
-}
-
-.video-event-detail .select-trigger {
-  height: 32px;
-}
-
-.video-event-detail .player {
-  height: 480px;
-}
-
-.video-event-detail .empty-table-replacement {
-  width: 100%;
-  text-align: center;
-}
