--- src/apps/EventDetail/Redesign/Player.css
@@ -1,121 +1,0 @@
-.player div {
-  margin: auto;
-}
-
-.player-control-panel {
-  height: 48px;
-  width: 100%;
-  background-color: var(--gray-200);
-  border-radius: 4px;
-  display: flex;
-  align-items: center;
-  box-sizing: border-box;
-}
-
-.player-control-panel .content {
-  margin: 0 10px;
-  display: flex;
-  width: 100%;
-  align-items: center;
-  justify-content: space-between;
-  gap: 10px;
-}
-
-.player-control-panel .content .position-label {
-  width: auto;
-  display: flex;
-  align-items: center;
-  justify-content: space-around;
-  gap: 4px;
-}
-
-.player-control-panel .content .position-label .position {
-  width: 70px;
-  margin: 0;
-}
-
-.player-control-panel .content .position-label .duration {
-  width: 72px;
-  font-weight: 600;
-}
-
-.player-control-panel .tooltip-content {
-  border-radius: 4px;
-  padding: 5px 8px;
-  font-size: 14px;
-  color: #000;
-  background-color: white;
-  border: 1px solid var(--gray-200);
-  user-select: none;
-  box-shadow: 0px 2px 4px 0px #00000026;
-}
-
-.player-control-panel .content .copy-button {
-  height: 24px;
-  width: 24px;
-  padding: 2px;
-  border-radius: 50%;
-}
-
-.player-control-panel .content .copy-button svg {
-  color: black;
-  height: 18px;
-  width: 18px;
-}
-
-.player-control-panel .content .seek-bar {
-  flex-grow: 1;
-}
-
-.player-control-panel .content .seek-bar .seek-bar-slider {
-  position: relative;
-  display: flex;
-  align-items: center;
-  user-select: none;
-  touch-action: none;
-  width: 100%;
-  height: 20px;
-}
-
-.player-control-panel .content .seek-bar .seek-bar-slider-track {
-  background-color: var(--gray-400);
-  position: relative;
-  flex-grow: 1;
-  border-radius: 9999px;
-  height: 3px;
-}
-
-.player-control-panel .content .seek-bar .seek-bar-slider-range {
-  position: absolute;
-  background-color: #000;
-  border-radius: 9999px;
-  height: 100%;
-}
-
-.player-control-panel .content .seek-bar .seek-bar-slider-thumb {
-  display: block;
-  width: 8px;
-  height: 8px;
-  background-color: #000;
-  border-radius: 10px;
-}
-
-.player-control-panel .content .seek-bar .seek-bar-slider-thumb:focus {
-  outline: none;
-  box-shadow: 0 0 0 2px var(--black-a8);
-}
-
-.player-control-panel .audio-button {
-  height: 36px;
-  width: 36px;
-  border-radius: 100%;
-}
-
-.player-control-panel .audio-button svg {
-  height: 24px;
-  width: 24px;
-}
-
-.player .offline {
-  font-style: italic;
-}
