+++ src/apps/EventDetail/SetManagementDropdown.tsx
