+++ src/apps/EventDetail/Tabs.css
