+++ src/apps/EventDetail/Tabs.tsx
