--- src/apps/EventDetail/index.ts
@@ -1,1 +1,0 @@
-export * from './EventDetail.tsx'
