+++ src/apps/EventDetail/index.tsx
