+++ src/apps/EventDetail/styles.css
