+++ src/pages/[lang]/projects/[projectSlug]/events/[eventUuid]/index.astro
@@ -1,5 +1,5 @@
 ---
-import EventDetail from '@apps/EventDetail/Redesign';
+import EventDetail from '@apps/EventDetail';
 import { getProject, parseSlug } from '@backend/projectHelpers';
 import { userInfo } from '@backend/userInfo';
 import { getTranslations } from '@i18n';
